backend/python/requirements.txt
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76
dlib==19.24.2 